+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - svg template for button
 - adding btn functionality and styles
 - add favicon, manifest, logo images to cache
+- move sources from `public` folder to `src'
 
 ## [0.4.0] - 2020-11-10
 ### Added